+++ packages/backend/src/context.ts
@@ -34,7 +34,6 @@ export async function createAppContext(io: IO) {
     submissionRepository,
     invitationRepository,
     entityRepository,
-    emailClient,
   )
   const schemaService = await createSchemaService(schemaRepository)
   const entityService = await createEntityService(
